services/book-service/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-filter>=23.0,<24.0
django-cors-headers>=4.0,<5.0
dj-database-url>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
gunicorn>=21.0,<22.0
